Concepto de apuesta y herencia de ronda
Se agrega clase apuesta para heredar en los diferentes juegos
