+++ casino-business/nbproject/private/private.xml
@@ -2,6 +2,11 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/C:/Users/WT/Documents/GitHub/Casino/casino-business/src/dominio/Apuesta.java</file>
+            <file>file:/C:/Users/WT/Documents/GitHub/Casino/casino-business/src/dominio/CasilleroRuleta.java</file>
+            <file>file:/C:/Users/WT/Documents/GitHub/Casino/casino-business/src/dominio/MesaRuleta.java</file>
+            <file>file:/C:/Users/WT/Documents/GitHub/Casino/casino-business/src/dominio/UnidadReceptoraDeJuego.java</file>
+        </group>
     </open-files>
 </project-private>
